Update label and chain schema validation
