+++ src/cashier_frontend/src/components/link-details/tip-link-asset-form.tsx
@@ -1,4 +1,4 @@
-import { ChangeEvent, FC, useEffect, useState } from "react";
+import { ChangeEvent, FC, useState } from "react";
 import { DefaultValues, SubmitHandler } from "react-hook-form";
 import {
     Form,
